add negative value bars to pulse chart, fix decimal investment input
- Pulse chart: negative values now show as red bars pointing downward
- Dynamic zero line appears only when negatives exist (15% space)
- Added color-danger-dark and color-danger-glow to color palette
- Investment per map now accepts decimal values (parseFloat + step=any)

+++ ui/js/settings.js
@@ -246,7 +246,7 @@ export async function saveSettings() {
     } else {
         settings.overlay_opacity = elements.settingOpacity.value / 100;
     }
-    settings.investment_per_map = parseInt(elements.settingInvestment.value) || 0;
+    settings.investment_per_map = parseFloat(elements.settingInvestment.value) || 0;
 
     // Save widget enabled states if in widget mode
     if (useWidgetOverlay && settings.widgets) {
